Fixed bug with setting password in email verification screen
