+++ backbone/src/main/java/org/researchstack/backbone/ui/step/layout/LoginStepLayout.java
@@ -70,7 +70,7 @@ public class LoginStepLayout extends ProfileStepLayout {
                     .compose(ObservableUtils.applyDefault());
 
             // Only gives a callback to response on success, the rest is handled by StepLayoutHelper
-            safePerformWithAlerts(login, this, response ->
+            StepLayoutHelper.safePerformWithAlerts(login, this, response ->
                     super.onNextClicked()
             );
         }
@@ -88,7 +88,7 @@ public class LoginStepLayout extends ProfileStepLayout {
                     .compose(ObservableUtils.applyDefault());
 
             // Only gives a callback to response on success, the rest is handled by StepLayoutHelper
-            safePerformWithAlerts(forgotPassword, this, response ->
+            StepLayoutHelper.safePerformWithAlerts(forgotPassword, this, response ->
                     showOkAlertDialog(response.getMessage())
             );
         }
